requirements.txt
chardet==4.0.0
cmake==3.18.4.post1
dlib==19.21.1
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.27.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
googledrivedownloader==0.4
googleDriveFileDownloader==1.2
imutils==0.5.2
numpy==1.15.4
opencv-python==4.1.0.25
dlib==19.17.0
pillow==8.1.0
aiogram==2.11.2
async-generator==1.10
async-timeout==3.0.1
aiohttp==3.7.3
